Storage isn't very clear about mirrored device management, add some short fixes on key points.
For example, when the docs say something "isn't possible" (like adding a disk), it may be possible for a mirror or single HDD but not for RAIDZ, and that can confuse users. Users with mirrors might try to find about managing them, which isn't covered, or think that "extending" covers "adding mirrors" which it doesn't.

I've tried to briefly improve this area. Please edit if it can be done better