Improve encryption header section
1) link to a more exact description about GELI key management

People using encryption (and managing its various keys) are likely to be the kind of people who will want exact details how keys are stored and managed. While they could Google GELI, that won't explain how %brand% uses and stores them, and is a lot to read through. However there's a nice and concise forum post which summarises it nicely and worth linking to for that info.....

2) Move "encryption performance" to its own small heading below the bullet list

Performance is always a question/concern when encrypting a data store (of any kind). This keeps the bullet list short and to the point, with general info+howto on %brand%'s encryption, and pulls the 1 or 2 bullets on performance to below the list, which probably flows better.