Clarify NUT UPS "user" account
Docs aren't clear if this is an internal client-to-daemon only account, an account that the user must create on %brand% or an account that has been created (or already exists) on the UPS itself and is used to authenticate and login for monitoring/control purposes.
As it's not clear, I hope I got this right! Please correct if not.