Clarify purpose of encryption and disposal of disks
section is unclear and contradictory. For example, states in one place that the primary purpose is secure wiping, in another it is physical theft, and nowhere states anything about temporary (illicit) removal + viewing while at rest. Editing 2 paragraphs to make this consistent and clear.

Also while about it, be clearer about disposal too, as it's much the same thing, and not well explained